# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: FixingApp Backend API
  description: |
    Backend MVP API for connecting employers with workers.
    
    This API provides endpoints for:
    - Authentication and user management
    - Job posting and management
    - Job applications
    - Worker verification
    - Reviews and ratings
    - Complaints handling
    - Admin functions
    
    **Authentication**: Most endpoints require a JWT Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fixingapp.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3000/api
    description: API base path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Not Found"
        message: "Route GET /api/invalid not found"
...
    Job:
      type: object
      properties:
        id:
          type: integer
          example: 1
        employerId:
          type: integer
          example: 1
        title:
          type: string
          example: "Fix leaking pipe"
        description:
          type: string
          example: "Need to fix a leaking pipe in kitchen"
        price:
          type: integer
          format: int64
          example: 150000
        address:
          type: string
          example: "123 Main St, City"
        requiredSkill:
          type: string
          example: "Plumbing"
        status:
          type: string
          enum: [CHUA_LAM, DANG_BAN_GIAO, DA_XONG]
          example: "CHUA_LAM"
        acceptedWorkerId:
          type: integer
          nullable: true
          example: null
        createdAt:
          type: integer
          format: int64
          example: 1702719041000
        updatedAt:
          type: integer
          format: int64
          example: 1702719041000
paths:
...
  /api/jobs:
    get:
      tags:
        - Jobs
      summary: List available jobs
      description: Get a list of open jobs (CHUA_LAM, DANG_BAN_GIAO status)
      operationId: listJobs
      parameters:
        - name: keyword
          in: query
          description: Search keyword
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category/skill
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: integer
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'

    post:
      tags:
        - Jobs
      summary: Create a new job
      description: Create a new job posting (Employer only)
      operationId: createJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - price
                - address
              properties:
                title:
                  type: string
                  example: "Fix leaking pipe"
                description:
                  type: string
                  example: "Need to fix a leaking pipe in kitchen"
                price:
                  type: integer
                  format: int64
                  example: 150000
                address:
                  type: string
                  example: "123 Main St, City"
                requiredSkill:
                  type: string
                  example: "Plumbing"
                images:
                  type: array
                  items:
                    type: string
                  example: ["/static/pipe1.jpg"]
            example:
              title: "Fix leaking pipe"
              description: "Need to fix a leaking pipe in kitchen"
              price: 150000
              address: "123 Main St, City"
              requiredSkill: "Plumbing"
              images: ["/static/pipe1.jpg"]
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{jobId}:
    get:
      tags:
        - Jobs
      summary: Get job details
      description: Get detailed information about a specific job
      operationId: getJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Jobs
      summary: Update job
      description: Update a job posting (only if status is CHUA_LAM and no worker accepted)
      operationId: updateJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: integer
                address:
                  type: string
                requiredSkill:
                  type: string
      responses:
        '200':
          description: Job updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request (e.g., cannot edit job with accepted worker)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Jobs
      summary: Delete job
      description: Delete a job posting (only if status is CHUA_LAM and no worker accepted)
      operationId: deleteJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Job deleted successfully
        '400':
          description: Bad request (cannot delete job with accepted worker)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/my:
    get:
      tags:
        - Jobs
      summary: Get my jobs
      description: Get all jobs posted by the authenticated employer
      operationId: getMyJobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/jobs/{jobId}/complete:
    post:
      tags:
        - Jobs
      summary: Complete a job
      description: Mark a job as completed (Job owner only)
      operationId: completeJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Job completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request (e.g., job not in DANG_BAN_GIAO status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{jobId}/reset:
    post:
      tags:
        - Jobs
      summary: Reset job status
      description: Reset job status from DANG_BAN_GIAO to CHUA_LAM (Job owner only)
      operationId: resetJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Job reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request (e.g., job not in DANG_BAN_GIAO status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not job owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...